kubevirt: move check-type selectors
