+++ frontend/packages/kubevirt-plugin/src/components/vm-disks/disk-row.tsx
@@ -10,7 +10,8 @@ import { DASH, dimensifyRow, getDeletetionTimestamp } from '@console/shared';
 import { TemplateModel } from '@console/internal/models';
 import { deleteDeviceModal, DeviceType } from '../modals/delete-device-modal';
 import { VMLikeEntityKind } from '../../types/vmLike';
-import { asVM, isVM, isVMI, isVMRunning } from '../../selectors/vm';
+import { asVM, isVMRunning } from '../../selectors/vm';
+import { isVM, isVMI } from '../../selectors/check-type';
 import { VirtualMachineModel } from '../../models';
 import { ValidationCell } from '../table/validation-cell';
 import { VMNicRowActionOpts } from '../vm-nics/types';
