+++ frontend/packages/kubevirt-plugin/src/components/vm-nics/nic-row.tsx
@@ -5,7 +5,8 @@ import { TemplateModel } from '@console/internal/models';
 import { DASH, dimensifyRow, getDeletetionTimestamp } from '@console/shared';
 import { deleteDeviceModal, DeviceType } from '../modals/delete-device-modal';
 import { VirtualMachineModel } from '../../models';
-import { asVM, isVM, isVMI, isVMRunning } from '../../selectors/vm';
+import { asVM, isVMRunning } from '../../selectors/vm';
+import { isVM, isVMI } from '../../selectors/check-type';
 import { VMLikeEntityKind } from '../../types/vmLike';
 import { nicModalEnhanced } from '../modals/nic-modal/nic-modal-enhanced';
 import { ValidationCell } from '../table/validation-cell';
