+++ frontend/packages/kubevirt-plugin/src/components/vms/vm-details.tsx
@@ -18,9 +18,9 @@ import { VirtualMachineInstanceModel, VirtualMachineModel } from '../../models';
 import { getServicesForVmi } from '../../selectors/service';
 import { VMResourceSummary, VMDetailsList, VMSchedulingList } from './vm-resource';
 import { VMTabProps } from './types';
-import { isVM, isVMI } from '../../selectors/vm/vmlike';
 import { getVMStatus } from '../../statuses/vm/vm-status';
 import { VMStatusBundle } from '../../statuses/vm/types';
+import { isVM, isVMI } from '../../selectors/check-type';
 
 export const VMDetailsFirehose: React.FC<VMTabProps> = ({
   obj: objProp,
