+++ frontend/packages/kubevirt-plugin/src/components/vms/vm.tsx
@@ -44,7 +44,7 @@ import { VMIKind, VMKind } from '../../types';
 import { buildOwnerReferenceForModel, getBasicID, getLoadedData } from '../../utils';
 import { getVMStatus } from '../../statuses/vm/vm-status';
 import { getVmiIpAddresses, getVMINodeName } from '../../selectors/vmi';
-import { isVM, isVMI } from '../../selectors/vm';
+import { isVMImport, isVM, isVMI } from '../../selectors/check-type';
 import { vmStatusFilter } from './table-filters';
 import { vmiMenuActions, vmImportMenuActions, vmMenuActions } from './menu-actions';
 import { VMILikeEntityKind } from '../../types/vmLike';
@@ -54,7 +54,6 @@ import { VMImportKind } from '../../types/vm-import/ovirt/vm-import';
 import { VMStatusBundle } from '../../statuses/vm/types';
 import { V1alpha1DataVolume } from '../../types/vm/disk/V1alpha1DataVolume';
 import { VMImportWrappper } from '../../k8s/wrapper/vm-import/vm-import-wrapper';
-import { isVMImport } from '../../selectors/check-type';
 import { getVMImportStatusAsVMStatus } from '../../statuses/vm-import/vm-import-status';
 import { V2VVMImportStatus } from '../../constants/v2v-import/ovirt/v2v-vm-import-status';
